extract build logic into a new class for clarity
